fix: resolve lint errors in ManifestEditor and tests
- Extract pathToLine to separate file to satisfy react-refresh rule
- Remove unused yamlLoad import
- Replace ref-in-render pattern with state-based diagnostics
- Fix test mock to use top-level React import instead of require()
- Remove unused waitFor import from test
